+++ R/theme_mprl.R
@@ -1,5 +1,5 @@
 #'
-#' @title Default theme used in all `ggstatsplot` package plots.
+#' @title Default theme used in all `ggstatsplot` package plots
 #' @name theme_mprl
 #' @aliases theme_ggstatsplot
 #' @author Indrajeet Patil